Make the :ins message cleaner with numbers
Show a list of who's in, interposed with commas, after the numbers. This
happens when (seq ins) and (not (nil? diner-count)) are both true. There
are also strings for each separately:

(seq ins) only: [what it was before]
mcd164 is *in*
mcd164 and dane are *in*

(not (nil? diner-count)) only: [technically unreachable]
3 of 12 people are *in*